requirements.txt
requests
pandas
google-generativeai
python-dotenv
schedule
beautifulsoup4
lxml
fastapi
uvicorn[standard]
jinja2
pytz
groq
streamlit
openai
